src/backend/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
blinker==1.7.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
idna==3.7
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kombu==5.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT>=2.0.0
python-dateutil==2.8.2
pytz==2024.1
pyzmq==25.1.2
redis==5.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
rq==1.15.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
yarg==0.1.9
